Add docs about strict separation of parameters
This addresses a question raised in #161 and is also kind-of related to #108.

The new section could also have been added to /parameters/, but i feel
it is more relevant here for users who are starting to learn about
subcommands. The behavior is totally irrelevant when writing a program
without subcommands, however sophisticated the usage of options and
arguments might be.