Fixed `invoked_subcommand`.
This is actually a very painful commit to make because it has to
change behavior.  On the bright side, this has actually been
broken from 3.1 and people have not noticed much yet, so maybe
it's not a big deal.  The docs also kinda left out the invoked
subcommands bit so I presume it is fine.

This fixes #205