Fix formatting of long argument lists
The text wrapper for the help text of argument lists is
configured with the width excluding the prefix.  If the prefix is
moderately long, this can result in the width being too short,
giving bizarre behaviour.

With this fix, the prefix is specified as the first line indent,
and the wrapper works correctly.