click.Path allow_dash broken
os.access('-', os.R_OK) is False (at least on my Mac), and so
even if click.Path(allow_dash=True) is used, then click.Path will
raise an error the check that '-' is readable. This adds a check
for is_dash on the readable check (since standard in is in fact
readable).