Make example work with setup tools entry points
Add handling for `ctx.obj is None` in an example command, with a comment
noting that the handling is there in case of invocations other than the
`if __name__ == '__main__'` block. This especially covers setuptools
entry points, but other usages as well.
Use `ctx.ensure_object` for clarity / as a current best-practice.

closes #572