_bashcomplete.py: Use -o nospace option instead of -o default
The -o default will cause the shell to complete paths in any
case that the program did not provide completions in COMPREPLY,
this is undesirable as it will result in completing filenames
when the next argument or option in context is not a filename.

We use nospace here so we can control whether to append a space
or not to the completions, this will allow us to complete paths
with python code and allow completing more dynamically (so that
after completing a subdirectory there is no space appended and
one can continue to complete the next subdir).