_bashcomplete.py: Use abstract complete() method for parameter completions
Instead of special casing the Choice type.

Note that this checks the user provided 'autocompletions' list in
advance of consulting the parameter class, this should allow users
to override the default completion list of a parameter.

Some minor adjustments made to append a space to builtin completion
lists of commands and option names so that this works seamlessly
with the new "complete -o nospace" approach.