Docs and examples: Updated for changed semantics of autocompletion
Slightly changed semantic for the autocompletion function which can
be specified in a decorator, now point out the semantic difference of
appending a space to a returned completion suggestion or not.

This should not be an API break since the autocompletion function
was added recently and has never been in a stable/official release.