test_bashcomplete.py: Modified for "complete -o nospace" changes.
After changing the bash wrapping script to invoke complete with the
nospace option, the decision to append a space and decide to start
completing the next program argument falls on click. So now all
results from get_choices() which are entirely "complete" and cannot
be further completed have an appended space.

Updated test case to expect this behavior from get_choices().