types.py: Handle Path completions at the root level gracefully
This was broken due to the way we split the input incomplete path
on the os separator token, when there is nothing to the left of
the last filesystem separator, we know we are trying to complete
from the root.

For windows, this should not happen since there will always
be a drive letter on the left of a valid terminating path separator.