Wrap stdout/stderr to avoid "Not enough space" in Python 2 on Windows 7
When you set the standard streams to binary you seem to only be able to
write up to some limit in a single call to write. So we wrap the streams
and write in chunks to avoid this.

A similar hack was also implemented in Python 3 to work around this. But
that hack doesn't seem to exist in Python 2.

This might have unintended side effects due to replacing the standard
stream objects.

Fixes #825