Fix issue #892
If the *choices* argument is not a list but a generator expression, it will eventually exhaust and lead to strange erros (e.g., the list of available choices looks empty, the order in which you pass choices is relevant). By wrapping it with `list()`, this should no longer be a problem.