feature: support customisation of "Commands:" section label.
The formatted text output currently defaults to a hardcoded `Commands` string.
This patch adds a kwarg that can override the formatted text outputed for a
given MultiCommand or Group.

So, for example, initialising `Group(section_label='Resources')` would print
`Resources:` as the formatted section label for a command group that represents
a list of resources with nested sub commands.