Ensure auto_envvar_prefix is always uppercased
Add a failing test and the modification to make it pass.

Given a lowercase auto_envvar_prefix, it's clearly supposed to be
converted to uppercase by the Context init. However, there's a slight
bug -- the prefix is uppercased, then reset to its prior value.

Test that setting a prefix of `test` results in `--arg` using
`TEST_ARG` (all uppercase).
Fix `Context.__init__` to do the right thing.