Add no_args_is_help to click.Command
The functionality for no_args_is_help, available on MultiCommands, can
easily be added on Commands as well. The result is not much more code,
and a more consistent interface between MultiCommands and Commands.
It's also nice to be able to write a click command `foo` and then just
hit `foo` for help, regardless of whether its a Group or Command.

closes #1156