Allow options to appear anywhere
When using commands, options must come before or after commands as dictated by
the heirarchy of the command tree. This makes thing easy to implement but burdensome
to the user. This patch allows options to appear before the command it applies
to or after a command even if it applies to a previous command. The last command
that can process the option wins. So "./script.py --help command" will print the
help of the script and "./script command --help" will print the help of the command.
If the script takes a flag "-v" and command takes a flag "-q", this command will
foward these flags appropriately: "./script -q command -v".