Fix to #1066 text wrapping
This commit addresses the issue with wrong breakage
when trying to wrap text with already given broken
lines. This was failing due to the text width, as we expandtabs
on every line. The width exceeds the limits and breaks. Does not
really break on the given (\n), rather it breaks when the width passes.
The fix in this commit checks if new line was introduced in a the description.
Then splits those into readable chunks and wrap the text based on each line instead
of wrapping up the text on all lines together which will end up reading the wrong
text width.