get default before processing value
This ensures the default value is processed like other values. Some
type converters were adjusted to accept values that are already the
correct type.

Use parameter source instead of value to determine if argument was
supplied on the command line during completion.

Add a parameter source for values from prompt.