fix argument with nargs=-1 and envvar (#1903)
A change in 3d3ea9c6 broke variadic arguments with an `envvar`
setting: the `process` method replaced `value` with `None`
uncoditionally, thus ignoring the values passed on the command line.
The intention of this was probably to replace the empty tuple (when no
arguments were passed) with `None` so that the value from the
environment would be picked later on. Adjust the condition accordingly
and add a test that covers this.