Extend Complex Application docs with Lazy Loading
Resolves #945

The goal of this example is to show how one might subclass Group, overriding
`get_command` and `list_commands` with some importlib usage and a dict, and produce a
viable lazy-loading solution.

One significant benefit of offering an example is that it provides space to explain how
and when `get_command` is invoked. i.e. Users might be surprised to realize that
subcommands are resolved during completion scenarios.

At the risk of diverging a small amount into what some would consider "general deferred
import semantics in python", the section closes out with an example of a `click.command`
decorated function which contains a deferred import. For codebases which have
significant at-import-time work (e.g. importing `requests` or `urllib3`), this strategy
is probably *more* useful than lazy loading of subcommands.

Other variations on this same solution are possible, e.g. using
`importlib.load_module(module_name, self.callback.__module__)` to handle imports
relative to the definition of the callback function. However, any further work in this
space is left as an exercise to readers of the doc.

Importantly, unlike some solutions already discussed in #945, nothing about this is
application-specific. The example LazyGroup implementation does not encode any knowledge
about the structure of the codebase in which it is used, which means that users can
copy-paste the example and expect it to work (within reason). Because this also
introduces the risk of reckless copy-paste by novice users, a warning is included in the
doc that is meant to point at a certain level of application maturity which should be
reached before this strategy can be safely applied.