Assert that all filenames that appear in messages can be printed
I've added an assertion helper with a large docstring to hopefully
ensure we don't regress on this point again, and have tried to add or
modify a test for every message which contains a filename.

@davidism suggested changing the error messages raised by click.Path()
from `{filename!r}` to `'{filename}'` in #2395, but I haven't done that
in this commit for two reasons:

1. It's slightly orthogonal, and whether we do it or not shouldn't be
   lumped in with the assertions about surrogates. It happens to be the
   case that using !r with a string containing surrogates will mask that
   things are broken since the resulting string will contain an escape
   sequence representing the surrogates.

   This is not particularly useful for the user, but it's at least
   something that can be printed.

2. Now that click.format_filename() does the right thing again and
   replaces errors with U+FFFD REPLACEMENT CHARACTER, using !r to quote
   works fine, e.g. 'foo�'. If the filename has really weird characters
   like \n, then I think printing it as \n rather than a literal newline
   is probably the right choice.

Some housekeeping:

I added a helper like the existing symlinks_supported one which was
previously imported from conftest. Importing from conftest is generally
discouraged. Additionally, conftest is only available to import as a
side effect of pytest's default import mode modifying sys.path.

I've changed the tests directory to a package so that relative imports
work and moved the helpers to a new tests/helpers.py file so that we're
not importing from conftest