Replace format for f-strings
Babel does not support f-strings, which prevents from localising a few
strings in the code, like "Usage: " and "Try".

The changes in this commit rewrites f-strings into the format syntax.
Please note some f-strings remain as they are not expected to be
translatable.

This commit has to deactivate the UP032 rule on a few files. There
seems to be an unreported bug in either Ruff or pyupgrade: The UP032
rule is not deactivated on multi-line commands, including multi-line
strings.