test: add two new tests for Parameter.__init__ in src/click/core.py
The function Parameter.__init__ does not have 100% coverage. This commit
adds two new test cases that improve the coverage. The first one tests
that a ValueError is raised when `nargs != self.type.arity`. The second
one tests that a ValueError is raised when `_check_iter(check_default)`
raises `TypeError`.

Fixes #2860
Recreation of #2861