Fix rendering when `prompt_suffix` is empty
This changes the implementation of the fixes in #1836 and #2093 to address
the problem that they created when `click.prompt` or `click.confirm` has a
`prompt_suffix` value which is empty.

This still fixes the problems reported in #665 and #2092 but does so without
modifying the prompt and adding a trailing space.

Previously, the problem was solved by trimming all trailing space characters
from the prompt and then adding a single trailing space to every prompt.
This was only visible if the last character in the prompt wasn't a space
character. In those cases a prompt like "This is my prompt" would render
as "This is my prompt ". Additionally, in the previous solution, if the
`err=True` argument was passed, that final trailing space character that
was added would be sent to STDOUT while the rest of the prompt would go
to STDERR. This also wasn't very apparent since it was a space character.

This different implementation uses the last character of the prompt to
work around #665 and #2092 instead of adding a space character. This fixes
most of the potentially confusing behavior. It does retain one confusing
behavior which is described above, where if you use `err=True` the last
character of the prompt is sent to STDOUT instead of STDERR. Previously,
this last character sent to STDOUT would always be a space character.
Now if the last character of the prompt is not a space character, that final
character will go to STDOUT.

For example if you call

```python
click.prompt("bar", prompt_suffix="", err=True)
```

The `ba` will go to STDERR and the `r` will go to STDOUT.
Previously `bar` would go to STDERR and an added ` ` would go to STDOUT.

This odd behavior, which is already present but just applies to a space
character, is worth accepting so that the `prompt_suffix` correctly renders
the suffix as it did in version 7.1.2 and prior.

This also adds unit tests for cases where `prompt_suffix` is empty.

This fixes #3019