Change autocomplete choise case sensitive test
The normalize function per default is mapping choices to lower case if
case sensitivity is set to False. As the normalize function can be
overwritten we have to take the value that is return by it also for
autocomplete, so we need to transform the auto complete value also into
lower case on default.