Fix: Bleeding Sentinel.UNSET to upstream users
Since release 8.3.0, the Sentinel.UNSET enum value has been bleeding to users of click where options have no default and the value is not required.

This commit contains a test which failed demonstrating the bleeding UNSET value and the fix in src/click/core.py to return to the previous behavior of returning None.