Fix default handling to defer UNSET normalization
When options are being parsed, multiple options may evaluate to have
defaults of `UNSET`, and some may evaluate to have explicitly set
defaults of `None`. The normalization of `UNSET -> None` is now
deferred until all options have been parsed, so that it is possible
to accurately evaluate and save the first non-UNSET default value.

A new test exercises this functionality by pointing two flags at the
same parameter declaration, and showing that their behavior is not order
sensitive when only one of the flags has a default.

resolves #3071